// frontend/src/client/models/response/browser-information.ts
export interface IHostLocation {
  name: string | null;
  version: string | null;
  versionNumber: string | null;
  os: string | null;
  country_code: string | null;
  country_name: string | null;
  city: string | null;
  postal: string | null;
  latitude: string | null;
  longitude: string | null;
  IPv4: string | null;
  state: string | null;
}